Move ConflictData to own files, fix namespace
