+++ GitCommands/GitCommands.csproj
@@ -87,6 +87,8 @@
     <Compile Include="GitRevisionInfoProvider.cs" />
     <Compile Include="Git\AheadBehindData.cs" />
     <Compile Include="Git\AheadBehindDataProvider.cs" />
+    <Compile Include="Git\ConflictData.cs" />
+    <Compile Include="Git\ConflictedFileData.cs" />
     <Compile Include="Git\ExecutableExtensions.cs" />
     <Compile Include="Git\ForcePushOptions.cs" />
     <Compile Include="Git\GitDescribeProvider.cs" />
